refactored - moved util function to `utils.ts`
