+++ src/minesweeper/utils.ts
@@ -1,4 +1,5 @@
 import {flatten, range, shuffle, take} from "lodash";
+import produce from "immer";
 
 import {BoardData, BoardDimension, CellCoordinates} from "./types";
 
@@ -36,4 +37,62 @@ export function blankBoardData({width, height}: BoardDimension): BoardData {
             isOpen: false,
         }))
     )
+}
+
+const EIGHT_WAYS_NEIGHBOURS: SimpleCoordinates[] = [
+    [-1, -1],
+    [-1, 0],
+    [-1, 1],
+    [0, -1],
+    [0, 1],
+    [1, -1],
+    [1, 0],
+    [1, 1],
+];
+
+export function initializedBoardData(
+    {oldBoard, clicked, numBomb}: { oldBoard: BoardData, clicked: CellCoordinates, numBomb: number }
+): BoardData {
+    const dimension: BoardDimension = getDimension(oldBoard)
+
+    // TODO: refactor this -> probably create a convenient function that takes both
+    const bombCandidates = getBombsList({takeCount: numBomb, dimension, clicked});
+
+    return produce(oldBoard, newBoard => {
+        bombCandidates.forEach(([row, col]) => {
+            newBoard[row][col].isBomb = true;
+            addCountsToNeighbour();
+
+            // TODO: extract this to global function for testability
+            function addCountsToNeighbour() {
+                EIGHT_WAYS_NEIGHBOURS
+                    .map(toExactCoordinates)
+                    .filter(keepValidCoordinates)
+                    .forEach(addCount)
+            }
+
+            function toExactCoordinates([drow, dcol]): SimpleCoordinates {
+                return [row + drow, col + dcol];
+            }
+
+            function keepValidCoordinates([exactRow, exactCol]) {
+                const isRowValid = (exactRow >= 0) && (exactRow < dimension.height)
+                const isColValid = (exactCol >= 0) && (exactCol < dimension.width)
+
+                return isRowValid && isColValid;
+            }
+
+            function addCount([exactRow, exactCol]) {
+                newBoard[exactRow][exactCol].count++;
+            }
+        })
+
+    })
+}
+
+function getDimension(boardData: BoardData): BoardDimension {
+    return {
+        width: boardData[0].length,
+        height: boardData.length
+    }
 }
